improvement: better CLI with subcommands
